Extract CBOR vars & funcs to new file cbor.go
